requirements.txt
gunicorn
flask
flask_cors
flask_sqlalchemy
flask_login
werkzeug==2.3.0
python-dotenv
pyjwt
psycopg2
psycopg2-binary
bs4
requests
python-qbittorrent
bencode.py
